Add overwrite functionality for renaming local
If the file or directory the user is attempting to rename it to already
exists, the user will be asked whether they would like to overwrite the
file. If the user inputs yes, then the renaming will proceed and will overwrite the
file. If the user inputs anything but yes, the rename method will
restart and they will be asked for the original file name and the new
name of the file.