Rewrite and reformat javadoc of each function
Rewrote and reformated javadocs so that all functions which do the same tasks for different variable inputs have the same javadoc structure. This was done so that is it is easier to recognize how functions relate and patterns in the code. Also added periods at the end of all full sentence comments per the Java Style Guide examples.