Rename variables in verification functions for uniqueness
Renamed the variables which are passed into all the verification functions so that they are unique to their purpose. Instead of having toVerify as the variable for both the username and host name functions, they are now usernameToVerify and hostnameToVerify. The password variable has been changed to passwordToVerify as well. This is done to help make a clear distinctions between different function variables and to reduce confusion.