Rename pattern variables specific to their input type
Renamed the pattern strings so that they are specific to their input type. Both the username pattern and hostname pattern variables were named userNamePattern which is confusing so two separate variables were made to distinguish the two. The capital case for Name in userNamePattern is not consistent with the rest of the code which keeps it lowercase so that was changed to "usernamePattern" instead of "userNamePattern".