Fix username input bug
Fixed a newly discovered username input verification bug which only has the user re-enter a username if invalid once and after than would accept an invalid username and move on to the password. The program now will ask to re-enter a username as many times as needed until the username is valid before moving on to entering the password.