Fix inconsistency in error handling output
The host name error handling output used to print out as a regular print statement instead of it being an error print statement just like the username and password methods. This has now been changed so that all error handling uses err.println for consistency and clarity when reading and using the program.